Revamped single-file plugins and main specification to be cleaner.
For single-file plugins, the contents of plugin.yml should be included
as a formatted comment at the top of the file. If 'main' is specified,
it refers to the name in the file to use as the plugin object or class.
Otherwise, 'Main', 'Plugin', and the 'quick' module are attempted.

For directory plugins, the main key should be a module name to import
optionally followed by ':' and a value name, interpreted as above.

Updated to Java 7 and Bukkit 1.7.9-R0.2 in pom.